split usb-host to multiple modules
